fix: separete external order into its own action
